Meet mrjob/Triadic Closure
This commit comprises exercises involving Facebook Social graph [data](http://snap.stanford.edu/data/egonets-Facebook.html) to find friends to suggest, friends list, and mutual friend counts. This dataset is less than a megabyte, so we can focus on the mechanics of MapReduce and ```mrjob``` and run it locally on our machines. If our data was much larger and stored in Amazon S3, we could easily run the same script on Amazon EMR.